+++ turbopack/crates/turbopack-ecmascript/src/lib.rs
@@ -15,7 +15,6 @@ pub mod chunk_group_files_asset;
 pub mod code_gen;
 mod errors;
 pub mod global_module_id_strategy;
-pub mod isolated_chunk;
 pub mod magic_identifier;
 pub mod manifest;
 pub mod minify;
@@ -32,6 +31,7 @@ pub mod tree_shake;
 pub mod typescript;
 pub mod utils;
 pub mod webpack;
+pub mod worker_chunk;
 
 use std::fmt::{Display, Formatter};
